stash:

+++ app/src/main/java/com/example/finalapp/data/model/dbEntity/FavouriteProductDbEntity.kt
@@ -11,7 +11,7 @@ class FavouriteProductDbEntity (
     @ColumnInfo(name = "description") val description: String?,
     @ColumnInfo(name = "iconUrl") val iconUrl: String?,
     @ColumnInfo(name = "price") val price: Int?,
-    @ColumnInfo(name = "rating") val rating: Int?,
+    @ColumnInfo(name = "rating") val rating: Float?,
     @ColumnInfo(name = "ratingsCount") val ratingsCount: Int?,
     @ColumnInfo(name = "commentsCount") val commentsCount: Int?
 )
